+++ src/matrix.js
@@ -107,7 +107,7 @@ export function summarizeFetch(resource, init) {
  * @returns {Promise<Object>}
  */
 export async function banUser(identity, roomId, userId, reason) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/ban`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/ban`, {
         method: 'POST',
         headers: {
             ...(identity.accessToken && {
@@ -130,7 +130,7 @@ export async function banUser(identity, roomId, userId, reason) {
  * @returns {Promise<Object>}
  */
 export async function createRoomAlias(identity, roomAlias, roomId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/directory/room/${encodeURIComponent(roomAlias)}`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/directory/room/${encodeURIComponent(roomAlias)}`, {
         method: 'PUT',
         headers: {
             ...(identity.accessToken && {
@@ -151,7 +151,7 @@ export async function createRoomAlias(identity, roomAlias, roomId) {
  * @returns {Promise<Object>}
  */
 export async function deleteRoomAlias(identity, roomAlias) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/directory/room/${encodeURIComponent(roomAlias)}`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/directory/room/${encodeURIComponent(roomAlias)}`, {
         method: 'DELETE',
         headers: {
             ...(identity.accessToken && {
@@ -166,30 +166,13 @@ export async function deleteRoomAlias(identity, roomAlias) {
 }
 
 /**
- * Forget about a room.
- * @param {Object} identity
- * @param {string} roomId
- * @returns {Promise<Object>}
- */
-export async function forgetRoom(identity, roomId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/forget`, {
-        method: 'POST',
-        headers: {
-            ...(identity.accessToken && {
-                Authorization: `Bearer ${identity.accessToken}`,
-            }),
-        },
-    });
-}
-
-/**
  * Gets a list of joined members of a room.
  * @param {Object} identity
  * @param {string} roomId
  * @returns {Promise<{joined: Record<string, Object>}>}
  */
 export async function getJoinedMembers(identity, roomId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/joined_members`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/joined_members`, {
         method: 'GET',
         headers: {
             ...(identity.accessToken && {
@@ -205,7 +188,7 @@ export async function getJoinedMembers(identity, roomId) {
  * @returns {Promise<{joined_rooms: string[]}>}
  */
 export async function getJoinedRooms(identity) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/joined_rooms`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/joined_rooms`, {
         method: 'GET',
         headers: {
             ...(identity.accessToken && {
@@ -240,7 +223,7 @@ export async function getMediaByRoom(identity, roomId) {
  * @returns {Promise<Object>}
  */
 export async function getMembers(identity, roomId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/members`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/members`, {
         method: 'GET',
         headers: {
             ...(identity.accessToken && {
@@ -259,7 +242,7 @@ export async function getMembers(identity, roomId) {
  * @returns {Promise<Object>}
  */
 export async function getState(identity, roomId, type, stateKey) {
-    let url = `${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/state`;
+    let url = `${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/state`;
     if (type) {
         url += `/${type}`;
     }
@@ -283,7 +266,7 @@ export async function getState(identity, roomId, type, stateKey) {
  * @returns {Promise<Object>}
  */
 export async function inviteUser(identity, roomId, userId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/invite`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/invite`, {
         method: 'POST',
         headers: {
             ...(identity.accessToken && {
@@ -298,23 +281,6 @@ export async function inviteUser(identity, roomId, userId) {
 }
 
 /**
- * Join a room.
- * @param {Object} identity
- * @param {string} roomId
- * @returns {Promise<Object>}
- */
-export async function joinRoom(identity, roomId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/join`, {
-        method: 'POST',
-        headers: {
-            ...(identity.accessToken && {
-                Authorization: `Bearer ${identity.accessToken}`,
-            }),
-        },
-    });
-}
-
-/**
  * Kick a user from a room.
  * @param {Object} identity
  * @param {string} roomId
@@ -323,7 +289,7 @@ export async function joinRoom(identity, roomId) {
  * @returns {Promise<Object>}
  */
 export async function kickUser(identity, roomId, userId, reason) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/kick`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/kick`, {
         method: 'POST',
         headers: {
             ...(identity.accessToken && {
@@ -339,30 +305,13 @@ export async function kickUser(identity, roomId, userId, reason) {
 }
 
 /**
- * Leave a room.
- * @param {Object} identity
- * @param {string} roomId
- * @returns {Promise<Object>}
- */
-export async function leaveRoom(identity, roomId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/leave`, {
-        method: 'POST',
-        headers: {
-            ...(identity.accessToken && {
-                Authorization: `Bearer ${identity.accessToken}`,
-            }),
-        },
-    });
-}
-
-/**
  * Resolves a room alias to a room ID.
  * @param {Object} identity
  * @param {string} roomAlias
  * @returns {{room_id: string, servers: string[]}}
  */
 export async function resolveAlias(identity, roomAlias) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/directory/room/${encodeURIComponent(roomAlias)}`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/directory/room/${encodeURIComponent(roomAlias)}`, {
         method: 'GET',
         headers: {
             ...(identity.accessToken && {
@@ -382,7 +331,7 @@ export async function resolveAlias(identity, roomAlias) {
  * @returns {Promise<Object>}
  */
 export async function sendEvent(identity, roomId, type, content, transactionId) {
-    let url = `${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/send/${encodeURIComponent(type)}`;
+    let url = `${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/send/${encodeURIComponent(type)}`;
     if (transactionId) {
         url += `/${transactionId}`;
     }
@@ -408,7 +357,7 @@ export async function sendEvent(identity, roomId, type, content, transactionId) 
  * @returns {Promise<Object>}
  */
 export async function setState(identity, roomId, type, stateKey, content) {
-    let url = `${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/state`;
+    let url = `${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/state`;
     if (type) {
         url += `/${type}`;
     }
@@ -435,7 +384,7 @@ export async function setState(identity, roomId, type, stateKey, content) {
  * @returns {Promise<Object>}
  */
 export async function unbanUser(identity, roomId, userId) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/rooms/${encodeURIComponent(roomId)}/unban`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/rooms/${encodeURIComponent(roomId)}/unban`, {
         method: 'POST',
         headers: {
             ...(identity.accessToken && {
@@ -457,7 +406,7 @@ export async function unbanUser(identity, roomId, userId) {
  * @returns {Promise<Object>}
  */
 export async function whoAmI(identity) {
-    return doRequest(`${identity.serverAddress}/_matrix/client/r0/account/whoami`, {
+    return doRequest(`${identity.serverAddress}/_matrix/client/v3/account/whoami`, {
         method: 'GET',
         headers: {
             ...(identity.accessToken && {
@@ -467,3 +416,12 @@ export async function whoAmI(identity) {
     });
 }
 
+/**
+ * Get a list of client features and versions supported by the server.
+ * @returns {Promise<Object>}
+ */
+export async function clientVersions() {
+    return doRequest(`${identity.serverAddress}/_matrix/client/versions`, {
+        method: 'GET',
+    });
+}